improved SendQueue config, fixed warning

+++ src/awds/settings.h
@@ -1,6 +1,7 @@
 #ifndef _SETTINGS_H__
 #define _SETTINGS_H__
 
+#include <stdint.h>
 
 static const int BEACON_INTERVAL = 1400;
 static const int TOPO_INTERVAL = 1000;
@@ -13,7 +14,13 @@ static const int TOPO_INTERVAL = 1000;
  */
 static const int NR_BEACON_TRIGGER_FAIL = 8;
 
-static const int QUEUE_LENGTH = 10;
+/**
+ * Length of a TX packet queue
+ *
+ * This limits the number of packets enqueued for transmission on the
+ * rawbasic device.
+ */
+static const size_t QUEUE_LENGTH = 10;
 
 #endif //SETTINGS_H__
 /* This stuff is for emacs
